# Implement iter::size_hint and ExactSizeIterator for Witness Iter