close https://github.com/rust-bitcoin/rust-bitcoin/issues/1050

I don't think we need to change the `collect` since it use the `size_hint()` lower bound to initially allocate

// with size_hint
// test blockdata::witness::benches::bench_big_witness_to_vec              ... bench:         313 ns/iter (+/- 13)
// test blockdata::witness::benches::bench_witness_to_vec                  ... bench:         204 ns/iter (+/- 11)

// without
// test blockdata::witness::benches::bench_big_witness_to_vec              ... bench:         489 ns/iter (+/- 28)
// test blockdata::witness::benches::bench_witness_to_vec                  ... bench:         221 ns/iter (+/- 102)


The reason why the small witness doesn't get big perf boost is because by default vec allocates 4 slots